debugging user controller

+++ src/utils/cloudinary.js
@@ -1,23 +1,29 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs"
+import dotenv from "dotenv";
+dotenv.config();
 
-  cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
-    });
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+});
 
-const uploadOnCloudinary=async (localFilePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath)return null
-        const response=await cloudinary.uploader.upload(localFilePath,{
-          resource_type:"auto"  
+        if (!localFilePath) return null
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto"
         })
-        console.log("File has been uploaded on cloudinary ",response.url);
+        fs.unlinkSync(localFilePath);
+        console.log("File has been uploaded on cloudinary ", response.url);
         return response
     } catch (error) {
-        fs.unlinkSync(localFilePath)
-        return null        
+        console.log("🔥 Cloudinary Upload Error:", error); // MUST
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        return null
     }
 }
